# Refactor and Optimize FA3 Code