<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
- Accelerate Spec Decoding + FA3 + CUDA Graph
- Easier to maintain

Follow up on https://github.com/sgl-project/sglang/pull/5050

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Benchmark for Llama 8B (Before and After)

| Iterations | GSM Latency on H200 | Send One on H200 |
|-------------------|-----------------------|-----------------------|
| Main Branch with Flash Infer           |  2118.780 token/s                    |   229.67 token/s               | 
| Main Branch  with FA3          |  2000.892 token/s                    |  235.28 token/s                 | 
| FA3 After commit: [21bf3fe](https://github.com/sgl-project/sglang/pull/5090/commits/21bf3fed77c73185f0161a247bc5d76d4e65c830)  and [eb77c88](https://github.com/sgl-project/sglang/pull/5090/commits/eb77c884a416c4bbd2b463497f0a39d295e72bf2)   | 2428.587 token/s               |  235.00 token/s |


GSM Accuracy is Accuracy: 0.775 with spec decoding.


## Sanity Check without Spec Decoding for Llama 8B
Accuracy: 0.775
Invalid: 0.000
Latency: 4.473 s

Accuracy: 0.775
Invalid: 0.000
Latency: 3.993 s
Output throughput: 4255.736 token/s

## Sanity Check with Deepseek V3 (Send One)
- Deepseek V3 with flashinfer + cuda graph: speed=41.59 token/s
- Deepseek V3 with flashinfer + disabled cuda graph: speed=4.60 token/s
- Deepseek V3 with flashinfer + cuda graph + spec decode: acc_length=2.59, speed=62.73 token/s
- Deepseek V3 with flashinfer + disabled cuda graph + spec decode:  acc_length=2.59, speed=11.09 token/s

- Deepseek V3 with fa3 + cuda graph: speed=44.93 token/s, result is good
- Deepseek V3 with fa3 + disabled cuda graph: speed=4.22 token/s token/s, result is good
- Deepseek V3 with fa3 + cuda graph + spec decode: acc_length=2.64, speed=66.66 token/s
- Deepseek V3 with fa3 + disabled cuda graph + spec decode: acc_length=2.69, speed=11.72 token/s 

## Sanity Check with Deepseek V3 (GSM8K)
- Deepseek V3 with flashinfer + cuda graph:
Accuracy: 0.960
Invalid: 0.000
Latency: 45.092 s
Output throughput: 439.656 token/s

- Deepseek V3 with flashinfer + disabled cuda graph:
Accuracy: 0.970
Invalid: 0.000
Latency: 154.613 s
Output throughput: 126.697 token/s

- Deepseek V3 with flashinfer + cuda graph + spec decode: 
Accuracy: 0.970
Invalid: 0.000
Latency: 42.325 s
Output throughput: 461.995 token/s

- Deepseek V3 with flashinfer + disabled cuda graph + spec decode: 
Accuracy: 0.965
Invalid: 0.000
Latency: 93.745 s
Output throughput: 209.964 token/s


- Deepseek V3 with fa3 + cuda graph: 
Accuracy: 0.965
Invalid: 0.000
Latency: 43.330 s
Output throughput: 463.840 token/s

- Deepseek V3 with fa3 + disabled cuda graph: 
Accuracy: 0.950
Invalid: 0.000
Latency: 178.272 s
Output throughput: 112.082 token/s


- Deepseek V3 with fa3 + cuda graph + spec decode: 
Accuracy: 0.955
Invalid: 0.000
Latency: 36.043 s
Output throughput: 550.533 token/s

- Deepseek V3 with fa3 + disabled cuda graph + spec decode:
Accuracy: 0.960
Invalid: 0.000
Latency: 134.789 s
Output throughput: 148.343 token/s

## Modifications

See all commits msg, they are clear

<!-- Describe the changes made in this PR. -->

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [x] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.